Fix #473
