+++ Environment_oM/Environment_oM.csproj
@@ -61,7 +61,7 @@
     <Compile Include="Gains\Enums\SimulationDayType.cs" />
     <Compile Include="Gains\Gain.cs" />
     <Compile Include="Gains\IGain.cs" />
-    <Compile Include="Materials\Enums\Function.cs" />
+    <Compile Include="MaterialFragments\Enums\Function.cs" />
     <Compile Include="Properties\Fragment\BuildingAnalyticalFragment.cs" />
     <Compile Include="Properties\Fragment\BuildingContextFragment.cs" />
     <Compile Include="Properties\Fragment\BuildingResultFragment.cs" />
@@ -82,14 +82,14 @@
     <Compile Include="Gains\Profile.cs" />
     <Compile Include="Gains\Thermostat.cs" />
     <Compile Include="IEnvironmentObject.cs" />
-    <Compile Include="Materials\Absorptance.cs" />
-    <Compile Include="Materials\Enums\AbsorptanceType.cs" />
-    <Compile Include="Materials\Enums\AbsorptanceUnit.cs" />
-    <Compile Include="Materials\Enums\Gas.cs" />
-    <Compile Include="Materials\Enums\Roughness.cs" />
-    <Compile Include="Materials\GasMaterial.cs" />
-    <Compile Include="Materials\IEnvironmentMaterial.cs" />
-    <Compile Include="Materials\SolidMaterial.cs" />
+    <Compile Include="MaterialFragments\Absorptance.cs" />
+    <Compile Include="MaterialFragments\Enums\AbsorptanceType.cs" />
+    <Compile Include="MaterialFragments\Enums\AbsorptanceUnit.cs" />
+    <Compile Include="MaterialFragments\Enums\Gas.cs" />
+    <Compile Include="MaterialFragments\Enums\Roughness.cs" />
+    <Compile Include="MaterialFragments\GasMaterial.cs" />
+    <Compile Include="MaterialFragments\IEnvironmentMaterial.cs" />
+    <Compile Include="MaterialFragments\SolidMaterial.cs" />
     <Compile Include="Properties\Fragment\IBHoMFragment.cs" />
     <Compile Include="Properties\LightingGain.cs" />
     <Compile Include="Properties\PeopleGain.cs" />
