+++ Environment_oM/MaterialFragments/Enums/AbsorptanceType.cs
@@ -20,14 +20,18 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-namespace BH.oM.Environment.Materials
+namespace BH.oM.Environment.MaterialFragments
 {
-    public enum Gas
+    public enum AbsorptanceType
     {
         Undefined,
-        Air,
-        Argon,
-        Krypton,
-        Xenon,
+        InfraredExternal,
+        InfraredInternal,
+        TotalExternal,
+        TotalInternal,
+        SolarExternal,
+        SolarInternal,
+        VisibleExternal,
+        VisibleInternal,
     }
 }
