+++ Environment_oM/MaterialFragments/Enums/AbsorptanceUnit.cs
@@ -20,16 +20,12 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-namespace BH.oM.Environment.Materials
+namespace BH.oM.Environment.MaterialFragments
 {
-    public enum Roughness
+    public enum AbsorptanceUnit
     {
         Undefined,
-        VeryRough,
-        MediumRough,
-        Rough,
-        Smooth,
-        MediumSmooth,
-        VerySmooth,
+        Fraction,
+        Percent,
     }
 }
